merging upstream changes to forked repo

+++ scripts/install-worker.sh
@@ -24,7 +24,8 @@ validate_env_set() {
 validate_env_set BINARY_BUCKET_NAME
 validate_env_set BINARY_BUCKET_REGION
 validate_env_set DOCKER_VERSION
-validate_env_set CNI_VERSION
+validate_env_set CONTAINERD_VERSION
+validate_env_set RUNC_VERSION
 validate_env_set CNI_PLUGIN_VERSION
 validate_env_set KUBERNETES_VERSION
 validate_env_set KUBERNETES_BUILD_DATE
@@ -63,7 +64,8 @@ sudo yum install -y \
     nfs-utils \
     socat \
     unzip \
-    wget
+    wget \
+    ipvsadm
 
 # Remove the ec2-net-utils package, if it's installed. This package interferes with the route setup on the instance.
 if yum list installed | grep ec2-net-utils; then sudo yum remove ec2-net-utils -y -q; fi
@@ -111,8 +113,23 @@ sudo yum install -y yum-utils device-mapper-persistent-data lvm2
 INSTALL_DOCKER="${INSTALL_DOCKER:-true}"
 if [[ "$INSTALL_DOCKER" == "true" ]]; then
     sudo amazon-linux-extras enable docker
-    sudo groupadd -fog 1950 docker && sudo useradd --gid 1950 docker
+    sudo groupadd -fog 1950 docker
+    sudo useradd --gid $(getent group docker | cut -d: -f3) docker
+
+    # install version lock to put a lock on dependecies
+    sudo yum install -y yum-plugin-versionlock
+
+    # install runc and lock version
+    sudo yum install -y runc-${RUNC_VERSION}
+    sudo yum versionlock runc-*
+
+    # install containerd and lock version
+    sudo yum install -y containerd-${CONTAINERD_VERSION}
+    sudo yum versionlock containerd-*
+
+    # install docker and lock version
     sudo yum install -y docker-${DOCKER_VERSION}*
+    sudo yum versionlock docker-*
     sudo usermod -aG docker $USER
 
     # Remove all options from sysconfig docker.
@@ -136,6 +153,7 @@ fi
 sudo mv $TEMPLATE_DIR/logrotate-kube-proxy /etc/logrotate.d/kube-proxy
 sudo mv $TEMPLATE_DIR/logrotate.conf /etc/logrotate.conf
 sudo chown root:root /etc/logrotate.d/kube-proxy
+sudo chown root:root /etc/logrotate.conf
 sudo mkdir -p /var/log/journal
 
 ################################################################################
@@ -151,6 +169,10 @@ echo "Downloading binaries from: s3://$BINARY_BUCKET_NAME"
 S3_DOMAIN="amazonaws.com"
 if [ "$BINARY_BUCKET_REGION" = "cn-north-1" ] || [ "$BINARY_BUCKET_REGION" = "cn-northwest-1" ]; then
     S3_DOMAIN="amazonaws.com.cn"
+elif [ "$BINARY_BUCKET_REGION" = "us-iso-east-1" ]; then
+    S3_DOMAIN="c2s.ic.gov"
+elif [ "$BINARY_BUCKET_REGION" = "us-isob-east-1" ]; then
+    S3_DOMAIN="sc2s.sgov.gov"
 fi
 S3_URL_BASE="https://$BINARY_BUCKET_NAME.s3.$BINARY_BUCKET_REGION.$S3_DOMAIN/$KUBERNETES_VERSION/$KUBERNETES_BUILD_DATE/bin/linux/$ARCH"
 S3_PATH="s3://$BINARY_BUCKET_NAME/$KUBERNETES_VERSION/$KUBERNETES_BUILD_DATE/bin/linux/$ARCH"
@@ -160,7 +182,7 @@ BINARIES=(
     aws-iam-authenticator
 )
 for binary in ${BINARIES[*]} ; do
-    if [[ ! -z "$AWS_ACCESS_KEY_ID" ]]; then
+    if [[ -n "$AWS_ACCESS_KEY_ID" ]]; then
         echo "AWS cli present - using it to copy binaries from s3."
         aws s3 cp --region $BINARY_BUCKET_REGION $S3_PATH/$binary .
         aws s3 cp --region $BINARY_BUCKET_REGION $S3_PATH/$binary.sha256 .
@@ -174,43 +196,31 @@ for binary in ${BINARIES[*]} ; do
     sudo mv $binary /usr/bin/
 done
 
+# Since CNI 0.7.0, all releases are done in the plugins repo.
+CNI_PLUGIN_FILENAME="cni-plugins-linux-${ARCH}-${CNI_PLUGIN_VERSION}"
+
 if [ "$PULL_CNI_FROM_GITHUB" = "true" ]; then
-    echo "Downloading CNI assets from Github"
-    wget https://github.com/containernetworking/cni/releases/download/${CNI_VERSION}/cni-${ARCH}-${CNI_VERSION}.tgz
-    wget https://github.com/containernetworking/cni/releases/download/${CNI_VERSION}/cni-${ARCH}-${CNI_VERSION}.tgz.sha512
-
-    wget https://github.com/containernetworking/plugins/releases/download/${CNI_PLUGIN_VERSION}/cni-plugins-${ARCH}-${CNI_PLUGIN_VERSION}.tgz
-    wget https://github.com/containernetworking/plugins/releases/download/${CNI_PLUGIN_VERSION}/cni-plugins-${ARCH}-${CNI_PLUGIN_VERSION}.tgz.sha512
-    sudo sha512sum -c cni-${ARCH}-${CNI_VERSION}.tgz.sha512
-    sudo sha512sum -c cni-plugins-${ARCH}-${CNI_PLUGIN_VERSION}.tgz.sha512
-    rm cni-${ARCH}-${CNI_VERSION}.tgz.sha512
-    rm cni-plugins-${ARCH}-${CNI_PLUGIN_VERSION}.tgz.sha512
+    echo "Downloading CNI plugins from Github"
+    wget "https://github.com/containernetworking/plugins/releases/download/${CNI_PLUGIN_VERSION}/${CNI_PLUGIN_FILENAME}.tgz"
+    wget "https://github.com/containernetworking/plugins/releases/download/${CNI_PLUGIN_VERSION}/${CNI_PLUGIN_FILENAME}.tgz.sha512"
+    sudo sha512sum -c "${CNI_PLUGIN_FILENAME}.tgz.sha512"
+    rm "${CNI_PLUGIN_FILENAME}.tgz.sha512"
 else
-    CNI_BINARIES=(
-            cni-${ARCH}-${CNI_VERSION}.tgz
-            cni-plugins-${ARCH}-${CNI_PLUGIN_VERSION}.tgz
-    )
-    for binary in ${CNI_BINARIES[*]} ; do
-        if [[ ! -z "$AWS_ACCESS_KEY_ID" ]]; then
-            echo "AWS cli present - using it to copy binaries from s3."
-            aws s3 cp --region $BINARY_BUCKET_REGION $S3_PATH/$binary .
-            aws s3 cp --region $BINARY_BUCKET_REGION $S3_PATH/$binary.sha256 .
-            sudo sha256sum -c $binary.sha256
-        else
-            echo "AWS cli missing - using wget to fetch cni binaries from s3. Note: This won't work for private bucket."
-            sudo wget $S3_URL_BASE/$binary
-            sudo wget $S3_URL_BASE/$binary.sha256
-        fi
-    done
+    if [[ -n "$AWS_ACCESS_KEY_ID" ]]; then
+        echo "AWS cli present - using it to copy binaries from s3."
+        aws s3 cp --region $BINARY_BUCKET_REGION $S3_PATH/${CNI_PLUGIN_FILENAME}.tgz .
+        aws s3 cp --region $BINARY_BUCKET_REGION $S3_PATH/${CNI_PLUGIN_FILENAME}.tgz.sha256 .
+    else
+        echo "AWS cli missing - using wget to fetch cni binaries from s3. Note: This won't work for private bucket."
+        sudo wget "$S3_URL_BASE/${CNI_PLUGIN_FILENAME}.tgz"
+        sudo wget "$S3_URL_BASE/${CNI_PLUGIN_FILENAME}.tgz.sha256"
+    fi
+    sudo sha256sum -c "${CNI_PLUGIN_FILENAME}.tgz.sha256"
 fi
-sudo tar -xvf cni-${ARCH}-${CNI_VERSION}.tgz -C /opt/cni/bin
-sudo tar -xvf cni-plugins-${ARCH}-${CNI_PLUGIN_VERSION}.tgz -C /opt/cni/bin
-rm cni-${ARCH}-${CNI_VERSION}.tgz
-rm cni-plugins-${ARCH}-${CNI_PLUGIN_VERSION}.tgz
+sudo tar -xvf "${CNI_PLUGIN_FILENAME}.tgz" -C /opt/cni/bin
+rm "${CNI_PLUGIN_FILENAME}.tgz"
 
-sudo rm *.sha256
-
-KUBERNETES_MINOR_VERSION=${KUBERNETES_VERSION%.*}
+sudo rm ./*.sha256
 
 sudo mkdir -p /etc/kubernetes/kubelet
 sudo mkdir -p /etc/systemd/system/kubelet.service.d
@@ -218,6 +228,12 @@ sudo mv $TEMPLATE_DIR/kubelet-kubeconfig /var/lib/kubelet/kubeconfig
 sudo chown root:root /var/lib/kubelet/kubeconfig
 sudo mv $TEMPLATE_DIR/kubelet.service /etc/systemd/system/kubelet.service
 sudo chown root:root /etc/systemd/system/kubelet.service
+# Inject CSIServiceAccountToken feature gate to kubelet config if kubernetes version starts with 1.20.
+# This is only injected for 1.20 since CSIServiceAccountToken will be moved to beta starting 1.21.
+if [[ $KUBERNETES_VERSION == "1.20"* ]]; then
+    KUBELET_CONFIG_WITH_CSI_SERVICE_ACCOUNT_TOKEN_ENABLED=$(cat $TEMPLATE_DIR/kubelet-config.json | jq '.featureGates += {CSIServiceAccountToken: true}')
+    echo $KUBELET_CONFIG_WITH_CSI_SERVICE_ACCOUNT_TOKEN_ENABLED > $TEMPLATE_DIR/kubelet-config.json
+fi
 sudo mv $TEMPLATE_DIR/kubelet-config.json /etc/kubernetes/kubelet/kubelet-config.json
 sudo chown root:root /etc/kubernetes/kubelet/kubelet-config.json
 
@@ -234,6 +250,18 @@ sudo mkdir -p /etc/eks
 sudo mv $TEMPLATE_DIR/eni-max-pods.txt /etc/eks/eni-max-pods.txt
 sudo mv $TEMPLATE_DIR/bootstrap.sh /etc/eks/bootstrap.sh
 sudo chmod +x /etc/eks/bootstrap.sh
+
+SONOBUOY_E2E_REGISTRY="${SONOBUOY_E2E_REGISTRY:-}"
+if [[ -n "$SONOBUOY_E2E_REGISTRY" ]]; then
+    sudo mv $TEMPLATE_DIR/sonobuoy-e2e-registry-config /etc/eks/sonobuoy-e2e-registry-config
+    sudo sed -i s,SONOBUOY_E2E_REGISTRY,$SONOBUOY_E2E_REGISTRY,g /etc/eks/sonobuoy-e2e-registry-config
+fi
+
+################################################################################
+### SSM Agent ##################################################################
+################################################################################
+
+sudo yum install -y amazon-ssm-agent
 
 ################################################################################
 ### AMI Metadata ###############################################################
@@ -260,33 +288,45 @@ kernel.panic_on_oops=1
 EOF
 
 ################################################################################
+### Setting up sysctl properties ###############################################
+################################################################################
+
+echo fs.inotify.max_user_watches=524288 | sudo tee -a /etc/sysctl.conf
+echo fs.inotify.max_user_instances=8192 | sudo tee -a /etc/sysctl.conf
+echo vm.max_map_count=524288 | sudo tee -a /etc/sysctl.conf
+
+
+################################################################################
 ### Cleanup ####################################################################
 ################################################################################
 
-# Clean up yum caches to reduce the image size
-sudo yum clean all
-sudo rm -rf \
-    $TEMPLATE_DIR  \
-    /var/cache/yum
+CLEANUP_IMAGE="${CLEANUP_IMAGE:-true}"
+if [[ "$CLEANUP_IMAGE" == "true" ]]; then
+    # Clean up yum caches to reduce the image size
+    sudo yum clean all
+    sudo rm -rf \
+        $TEMPLATE_DIR  \
+        /var/cache/yum
 
-# Clean up files to reduce confusion during debug
-sudo rm -rf \
-    /etc/hostname \
-    /etc/machine-id \
-    /etc/resolv.conf \
-    /etc/ssh/ssh_host* \
-    /home/ec2-user/.ssh/authorized_keys \
-    /root/.ssh/authorized_keys \
-    /var/lib/cloud/data \
-    /var/lib/cloud/instance \
-    /var/lib/cloud/instances \
-    /var/lib/cloud/sem \
-    /var/lib/dhclient/* \
-    /var/lib/dhcp/dhclient.* \
-    /var/lib/yum/history \
-    /var/log/cloud-init-output.log \
-    /var/log/cloud-init.log \
-    /var/log/secure \
-    /var/log/wtmp
+    # Clean up files to reduce confusion during debug
+    sudo rm -rf \
+        /etc/hostname \
+        /etc/machine-id \
+        /etc/resolv.conf \
+        /etc/ssh/ssh_host* \
+        /home/ec2-user/.ssh/authorized_keys \
+        /root/.ssh/authorized_keys \
+        /var/lib/cloud/data \
+        /var/lib/cloud/instance \
+        /var/lib/cloud/instances \
+        /var/lib/cloud/sem \
+        /var/lib/dhclient/* \
+        /var/lib/dhcp/dhclient.* \
+        /var/lib/yum/history \
+        /var/log/cloud-init-output.log \
+        /var/log/cloud-init.log \
+        /var/log/secure \
+        /var/log/wtmp
+fi
 
 sudo touch /etc/machine-id
